renamed new download_packages functions files to fit with scopesim
